refactor(test): split config test suites and fix lint hygiene
