+++ test/package-types.test.ts
@@ -48,6 +48,7 @@ describe("published package types", () => {
         "NodeNext",
         "--target",
         "ES2022",
+        "--ignoreConfig",
         "--skipLibCheck",
         entryPath,
       ],
@@ -102,6 +103,7 @@ describe("published package types", () => {
         "NodeNext",
         "--target",
         "ES2022",
+        "--ignoreConfig",
         "--skipLibCheck",
         entryPath,
       ],
